backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langgraph>=0.0.62
langchain>=0.1.0
langchain-google-genai>=0.0.6
google-generativeai>=0.3.2
pydantic>=2.5.0
python-multipart>=0.0.6
requests>=2.31.0
pillow>=10.1.0
opencv-python>=4.8.1.78
yt-dlp>=2023.12.30
numpy>=1.24.3
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.25.2